plant-disease-detection/requirements.txt
Flask
numpy
Pillow
tensorflow-cpu